[TACHYON-1199] Extract start method into super class
